Disable output of dev-mode web-resources and web-resource conditions until the feature is re-implemented.
The extraction from our apps assumes that there will be certain Java-based Condition implementations available in the product or plugin.
This is not going to be the case for third party plugins, until such time as there is a common maven plugin that ships them.

This feature will re-appear when this webpack plugin's API can be updated to allow pluggability for conditions in generated web-resources.