Use the DllModule for any external web-resource inclusions.
The ExternalModule class expects to output a variable that can be depended upon and used within the source of a module. When we're importing web-resources, we're doing so because we want to declare dependencies on legacy code. We don't get anything back from a web-resource, and we don't want to affect the compiled module. Using a DllModule makes more sense because it won't affect the output in the bundle, and will let us focus on affecting the web-resource XML we generate.