Replace use of 'wrc!' and 'wr!' with 'wr-dependency!', so that it's more obvious what it does (adds a <dependency> block on a webresource to the generated XML) and doesn't confuse things with what WRM.require does today.
The 'wrc!' and 'wr!' loader prefixes are used at product runtime by the WRM to load various stuff asynchronously. Here, we're using the flags to compile the dependency declaration ahead of time. It's a different use-case, so deserves a different prefix.

Initially I thought 'webresource!' would be clear enough, but it doesn't imply what it's going to do about the web-resource key once it's added. Tim's suggestion of 'wr-dependency!' makes much more sense, as it both prefixes the loader to indicate it's a web-resource concern, and declares what it'll generate in the XML.