Explicitly export nothing when including a web-resource dependency via a JavaScript import statement. (fixes #11)
This usage of the DllModule is a slight perversion of its original intent. Because we're not actually linking any external bundles, webpack doesn't provide the __webpack_require__ function to the generated chunk for our web-resource imports. This results in a '__webpack_require__ is not defined' error at runtime.

We don't actually care about pulling in external values right now, and we don't want to expose any value for these chunks, so it's fine to export undefined.