Fix the external resource inclusion so that more than one will be added at a time.
Previously, they were being culled because the resource's "name" as a DLL Module was always 'null', and DLL Modules are de-duplicated by their names. They were also outputting an extraneous <dependency> because we were extending ProvidedDllModule, which is there to provide external web-resource dependencies. In short, the class names are confused, and need refactoring too.