Issue #55: add locationPrefix configuration option
Currently to set a prefix for web-resource location attribute one needs
to set webpack output.path to something that starts with 'target/classes'
which imposes maven-like project structure.

With locationPrefix option it is possible to set it any value and override
previous behavior of the plugin extracting it from output.path. This
makes possible to use it with any project structure.

When locationPrefix is not provided plugin will fallback to previous
behavior of extracting it from output.path property.