refactor: base contexts are configurable.
Rather than unconditionally providing the WRM dependencies, we instead only add them when the webpack bundle is affected and makes use of the WRM's APIs.

This is a bit better than things were before, but it isn't a general-case solution. This does not add the WRM API requirements to only the web-resource blocks that need them; it still adds them to everything. A correct solution would only add dependencies on WRM APIs to the web-resource blocks that actually call them.