fix: only add the dependency on WRM.require if it is needed by a chunk.
The requireEnsure hook is called when the compilation is creating a chunk that loads async chunks. As such, it makes sense to tie these things together.

This is getting even closer to what we want. Some tests are failing, but that might be because our expectations weren't correct at the time of writing the test.