fix: assets served on the public path will get cache headers in Atlassian Server apps.
Server applications implement a UrlRewriteFilter that searches for URLs with a regex like `/s/(.*?)/_/download/.*`,
which is the signifier that cache headers should be added to whatever is being served over that URL.
The code that adds the headers lives in the atlassian-plugins-webresource-common module. The code that calls it lives in each product.

It doesn't matter what value we add in-between the /s/ and /_/ URL parts, only that if the contents of a bundled asset changes, the value changes too.
This is the expected cache-busting mechanism in Atlassian Server apps.