refactor: move code closer to where it is used and needed.
all the code related to generating QUnit dependencies is now isolated to the QUnit plugin.

the web-resource generator module does one thing: render web-resource XML strings.

the provided dependency code is in a single module to make it easier to find and change.

renamed functions to be more consistent... e.g., anything starting with 'render' is expected to return a string.

removed some unneeded abstractions... e.g., createResourceDescriptors was a wrapper around a map operation that was improperly handled anyway.